+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Filter out app start with more than 60s ([#2303](https://github.com/getsentry/sentry-react-native/pull/2303))
+
 ## 4.0.0
 
 - Bump Sentry Cocoa 7.18.0 ([#2303](https://github.com/getsentry/sentry-react-native/pull/2303))